+++ packages/ui/src/alert-dialog/popup.tsx
@@ -30,6 +30,7 @@ const Popup = forwardRef< HTMLDivElement, PopupProps >(
 			children,
 			confirmButtonText = __( 'OK' ),
 			cancelButtonText = __( 'Cancel' ),
+			portalClassName,
 			...props
 		},
 		ref
@@ -45,7 +46,10 @@ const Popup = forwardRef< HTMLDivElement, PopupProps >(
 		const buttonsDisabled = phase !== 'idle' || undefined;
 
 		return (
-			<_AlertDialog.Portal container={ container }>
+			<_AlertDialog.Portal
+				container={ container }
+				className={ portalClassName }
+			>
 				<_AlertDialog.Backdrop className={ dialogStyles.backdrop } />
 				<ThemeProvider>
 					<_AlertDialog.Popup