+++ packages/ui/src/dialog/popup.tsx
@@ -28,6 +28,7 @@ const Popup = forwardRef< HTMLDivElement, PopupProps >( function DialogPopup(
 		size = 'medium',
 		initialFocus,
 		finalFocus,
+		portalClassName,
 		children,
 		...props
 	},
@@ -40,7 +41,7 @@ const Popup = forwardRef< HTMLDivElement, PopupProps >( function DialogPopup(
 	const mergedRef = useMergeRefs( [ ref, popupRef ] );
 
 	return (
-		<_Dialog.Portal container={ container }>
+		<_Dialog.Portal container={ container } className={ portalClassName }>
 			<_Dialog.Backdrop className={ styles.backdrop } />
 			<ThemeProvider>
 				<_Dialog.Popup